Update bytes features with lasted bytes version 1.1.0 & add edition 2021
