+++ Cargo.toml
@@ -13,8 +13,9 @@ edition = "2021"
 
 [features]
 
-[dependencies]
-bytes = { version = "0.5", optional = true }
+[dependencies.bytes]
+version = "1.1.0"
+optional = true
 
 [dev-dependencies]
 vecio = "0.1.0"
